酒店管理测试完成，房间管理开发完成

+++ hn-platform/hn-travel-persist/src/main/java/hn/travel/persist/service/hotel/HotelRoomService.java
@@ -3,17 +3,22 @@
  */
 package hn.travel.persist.service.hotel;
 
+import hn.travel.persist.entity.BlobData;
 import hn.travel.persist.entity.HotelRoom;
 import hn.travel.persist.repository.BlobDataDao;
 import hn.travel.persist.repository.HotelRoomDao;
 
 import java.util.ArrayList;
 import java.util.Arrays;
+import java.util.Date;
 import java.util.List;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.data.domain.Page;
+import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Component;
 import org.springframework.transaction.annotation.Transactional;
+import org.springframework.util.StringUtils;
 
 /**
  * @author XFZP
@@ -49,20 +54,98 @@ public class HotelRoomService {
 
 	@Transactional
 	public void delete(Iterable<? extends HotelRoom> hotelRooms) {
-		dao.delete(hotelRooms);
+		if (hotelRooms == null)
+			return;
 
 		List<Long> idList = new ArrayList<Long>();
+		List<Long> bdIdList = new ArrayList<Long>();
 		for (HotelRoom hr : hotelRooms) {
 			if (hr.getInfoId() != null)
-				blobDataDao.delete(hr.getInfoId());
+				bdIdList.add(hr.getInfoId());
 			if (hr.getCostId() != null)
-				blobDataDao.delete(hr.getCostId());
+				bdIdList.add(hr.getCostId());
 			if (hr.getNoticeId() != null)
-				blobDataDao.delete(hr.getNoticeId());
+				bdIdList.add(hr.getNoticeId());
 			idList.add(hr.getId());
 		}
+		if (idList.size() == 0)
+			return;
 
 		roomKindSrv.deleteByRoomIds(idList.toArray(new Long[idList.size()]));
+
+		dao.delete(hotelRooms);
+
+		if (bdIdList.size() > 0) {
+			Iterable<BlobData> bds = blobDataDao.findAll(bdIdList);
+			blobDataDao.delete(bds);
+		}
+	}
+
+	public Page<HotelRoom> page(Long hotelId, Pageable pageable) {
+		return dao.findByHotelId(hotelId, pageable);
+	}
+
+	public HotelRoom getDetail(Long id) {
+		HotelRoom hr = dao.findOne(id);
+		if (hr != null) {
+			List<Long> ids = new ArrayList<Long>(3);
+			if (hr.getInfoId() != null)
+				ids.add(hr.getInfoId());
+			if (hr.getCostId() != null)
+				ids.add(hr.getCostId());
+			if (hr.getNoticeId() != null)
+				ids.add(hr.getNoticeId());
+
+			if (ids.size() > 0) {
+				for (BlobData bd : blobDataDao.findAll(ids)) {
+					if (bd.getId().equals(hr.getInfoId())) {
+						hr.setInfo(bd.getData());
+					} else if (bd.getId().equals(hr.getCostId())) {
+						hr.setCost(bd.getData());
+					} else if (bd.getId().equals(hr.getNoticeId())) {
+						hr.setNotice(bd.getData());
+					}
+				}
+			}
+		}
+		return hr;
+	}
+
+	@Transactional
+	public HotelRoom save(HotelRoom hr) {
+		if (StringUtils.hasLength(hr.getInfo())) {
+			BlobData bd = blobDataDao.save(new BlobData(hr.getInfoId(), hr
+					.getInfo()));
+			hr.setInfoId(bd.getId());
+		} else if (hr.getInfoId() != null) {
+			blobDataDao.delete(hr.getInfoId());
+			hr.setInfoId(null);
+		}
+
+		if (StringUtils.hasLength(hr.getCost())) {
+			BlobData bd = blobDataDao.save(new BlobData(hr.getCostId(), hr
+					.getCost()));
+			hr.setCostId(bd.getId());
+		} else if (hr.getCostId() != null) {
+			blobDataDao.delete(hr.getCostId());
+			hr.setCostId(null);
+		}
+
+		if (StringUtils.hasLength(hr.getNotice())) {
+			BlobData bd = blobDataDao.save(new BlobData(hr.getNoticeId(), hr
+					.getNotice()));
+			hr.setNoticeId(bd.getId());
+		} else if (hr.getNoticeId() != null) {
+			blobDataDao.delete(hr.getNoticeId());
+			hr.setNoticeId(null);
+		}
+
+		if (hr.getCreateTime() == null)
+			hr.setCreateTime(new Date());
+		hr.setUpdateTime(new Date());
+
+		dao.save(hr);
+		return hr;
 	}
 
 }
